requirements.txt
blinker==1.5
certifi==2022.12.7
duckdb==0.7.1
kafka-python==2.0.2
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.3
sentry-sdk==1.17.0
six==1.16.0
urllib3==1.26.15
